+++ Sources/Cider/Client/CiderClient.swift
@@ -56,9 +56,9 @@ public struct CiderClient {
        - types: The `MediaType`s to limit the search to.
        - completion: The completion handler to call with the results of the search.
      */
-    public func search(term: String, limit: Int? = nil, offset: Int? = nil, types: [MediaType]? = nil, completion: ((ResponseRoot<SearchResults>?, Error?) -> Void)?) {
+    public func search(term: String, limit: Int? = nil, offset: Int? = nil, types: [MediaType]? = nil, completion: ((SearchResults?, Error?) -> Void)?) {
         let request = urlBuilder.searchRequest(term: term, limit: limit, offset: offset, types: types)
-        fetch(request, completion: completion)
+        fetch(request) { (results: ResponseRoot<SearchResults>?, error) in completion?(results?.results, error) }
     }
 
     /**
@@ -70,9 +70,9 @@ public struct CiderClient {
        - types: The `MediaType`s to limit the search to.
        - completion: The completion handler to call with the results of the search hints.
      */
-    public func searchHints(term: String, limit: Int? = nil, types: [MediaType]? = nil, completion: ((ResponseRoot<SearchHints>?, Error?) -> Void)?) {
+    public func searchHints(term: String, limit: Int? = nil, types: [MediaType]? = nil, completion: ((SearchHints?, Error?) -> Void)?) {
         let request = urlBuilder.searchHintsRequest(term: term, limit: limit, types: types)
-        fetch(request, completion: completion)
+        fetch(request) { (results: ResponseRoot<SearchHints>?, error) in completion?(results?.results, error) }
     }
 
     // MARK: Lookup
@@ -85,9 +85,9 @@ public struct CiderClient {
        - include: The relationships to include in the lookup.
        - completion: The handler to call with the results.
      */
-    public func artist(id: String, include: [Include]? = nil, completion: ((ResponseRoot<Artist>?, Error?) -> Void)?) {
+    public func artist(id: String, include: [Include]? = nil, completion: ((Artist?, Error?) -> Void)?) {
         let request = urlBuilder.fetchRequest(mediaType: .artists, id: id, include: include)
-        fetch(request, completion: completion)
+        fetch(request) { (results: ResponseRoot<Artist>?, error) in completion?(results?.data?.first, error) }
     }
 
     /**
@@ -98,9 +98,9 @@ public struct CiderClient {
        - include: The relationships to include in the lookup.
        - completion: The handler to call with the results.
      */
-    public func album(id: String, include: [Include]? = nil, completion: ((ResponseRoot<Album>?, Error?) -> Void)?) {
+    public func album(id: String, include: [Include]? = nil, completion: ((Album?, Error?) -> Void)?) {
         let request = urlBuilder.fetchRequest(mediaType: .albums, id: id, include: include)
-        fetch(request, completion: completion)
+        fetch(request) { (results: ResponseRoot<Album>?, error) in completion?(results?.data?.first, error) }
     }
 
     /**
@@ -111,9 +111,9 @@ public struct CiderClient {
        - include: The relationships to include in the lookup.
        - completion: The handler to call with the results.
      */
-    public func song(id: String, include: [Include]? = nil, completion: ((ResponseRoot<Track>?, Error?) -> Void)?) {
+    public func song(id: String, include: [Include]? = nil, completion: ((Track?, Error?) -> Void)?) {
         let request = urlBuilder.fetchRequest(mediaType: .songs, id: id, include: include)
-        fetch(request, completion: completion)
+        fetch(request) { (results: ResponseRoot<Track>?, error) in completion?(results?.data?.first, error) }
     }
 
     /**
@@ -124,9 +124,9 @@ public struct CiderClient {
        - include: The relationships to include in the lookup.
        - completion: The handler to call with the results.
      */
-    public func playlist(id: String, include: [Include]? = nil, completion: ((ResponseRoot<Playlist>?, Error?) -> Void)?) {
+    public func playlist(id: String, include: [Include]? = nil, completion: ((Playlist?, Error?) -> Void)?) {
         let request = urlBuilder.fetchRequest(mediaType: .playlists, id: id, include: include)
-        fetch(request, completion: completion)
+        fetch(request) { (results: ResponseRoot<Playlist>?, error) in completion?(results?.data?.first, error) }
     }
 
     /**
@@ -137,9 +137,9 @@ public struct CiderClient {
        - include: The relationships to include in the lookup.
        - completion: The handler to call with the results.
      */
-    public func musicVideo(id: String, include: [Include]? = nil, completion: ((ResponseRoot<MusicVideo>?, Error?) -> Void)?) {
+    public func musicVideo(id: String, include: [Include]? = nil, completion: ((MusicVideo?, Error?) -> Void)?) {
         let request = urlBuilder.fetchRequest(mediaType: .musicVideos, id: id, include: include)
-        fetch(request, completion: completion)
+        fetch(request) { (results: ResponseRoot<MusicVideo>?, error) in completion?(results?.data?.first, error) }
     }
 
     /**
@@ -150,9 +150,9 @@ public struct CiderClient {
        - include: The relationships to include in the lookup.
        - completion: The handler to call with the results.
      */
-    public func curator(id: String, include: [Include]? = nil, completion: ((ResponseRoot<Curator>?, Error?) -> Void)?) {
+    public func curator(id: String, include: [Include]? = nil, completion: ((Curator?, Error?) -> Void)?) {
         let request = urlBuilder.fetchRequest(mediaType: .curators, id: id, include: include)
-        fetch(request, completion: completion)
+        fetch(request) { (results: ResponseRoot<Curator>?, error) in completion?(results?.data?.first, error) }
     }
 
     // MARK: Relationships
@@ -166,10 +166,10 @@ public struct CiderClient {
        - offset: The offset to use for pagination.
        - completion: The handler to call with the results.
      */
-    public func get<T>(related: Relationship<T>, limit: Int? = nil, offset: Int? = nil, completion: ((ResponseRoot<T>?, Error?) -> Void)?) {
+    public func get<T>(related: Relationship<T>, limit: Int? = nil, offset: Int? = nil, completion: (([T]?, Error?) -> Void)?) {
         let path = related.href
         let request = urlBuilder.relationshipRequest(path: path, limit: limit, offset: offset)
-        fetch(request, completion: completion)
+        fetch(request) { (results: ResponseRoot<T>?, error) in completion?(results?.data, error) }
     }
 
     // MARK: Helpers