+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 ##### Breaking
 
+* None
+
+##### Enhancements
+
+* None
+
+##### Bug Fixes
+
+* None
+
+## 0.10.0
+
+##### Breaking
+
 * Most properties of `Artwork` are now optional since music video artwork does
   not contain them.
 * Return one level more specific instead of `ResponseRoot` for most requests.